+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.29.14 Unreleased] - 2021-10-06
+### Fixed
+- Fixed copyFilesTo from .shadow to .deleted for Windows
+
 ## [2.29.13] - 2021-10-06
 ### Changed
 - initHandler class introduced
@@ -25,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.29.9] - 2021-09-28
 ### Changed
-- copyFilesTo reverted change for formmated Path
+- copyFilesTo reverted change for formatted Path
 
 ## [2.29.8] - 2021-09-28
 ### Changed